+++ api/candidates/models/__init__.py
@@ -1,0 +1,4 @@
+from .candidates import Candidates
+
+
+__all__ = [Candidates]
